requirements.txt
neo4j
python-dotenv
langchain-neo4j
langchain-community